refactor: remove plugins and astro project, move markdown generation to scripts
