+++ docs/.gitignore
@@ -4,9 +4,6 @@ src/content/tmp/
 src/content/docs/libs/
 src/content/docs/changelog.md
 
-# generated types
-.astro/
-
 # dependencies
 node_modules/
