+++ docs/package.json
@@ -4,14 +4,12 @@
   "type": "module",
   "version": "0.0.1",
   "scripts": {
-    "astro": "astro",
-    "start": "tsc -b tsconfig.libs.json && astro dev",
-    "build": "tsc -b tsconfig.libs.json && astro build",
-    "preview": "astro preview"
+    "build": "pnpm dlx tsx scripts/generate-libs-docs.ts"
   },
   "devDependencies": {
-    "@astrojs/starlight": "^0.34.4",
-    "@dfinity/starlight": "workspace:*",
-    "astro": "^5.13.2"
+    "@tsconfig/node-ts": "^23.6.1",
+    "@tsconfig/node22": "^22.0.2",
+    "typedoc-plugin-frontmatter": "^1.3.0",
+    "typedoc-plugin-markdown": "^4.8.1"
   }
 }
