--- docs/plugins/README.md
@@ -1,9 +1,0 @@
-# DFINITY Starlight Plugins
-
-## Overview
-
-Starlight plugins for use in DFINITY documentation websites.
-
-## License
-
-This project is licensed under the [Apache-2.0 License](LICENSE).
