--- docs/plugins/package.json
@@ -1,52 +1,0 @@
-{
-  "name": "@dfinity/starlight",
-  "private": true,
-  "version": "0.0.0",
-  "author": "DFINITY Stiftung <sdk@dfinity.org>",
-  "license": "Apache-2.0",
-  "description": "DFINITY plugins for Starlight.",
-  "homepage": "https://github.com/dfinity/starlight-plugins/README.md",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/dfinity/starlight-plugins.git"
-  },
-  "bugs": {
-    "url": "https://github.com/dfinity/starlight-plugins/issues"
-  },
-  "keywords": [
-    "dfinity",
-    "starlight",
-    "plugins",
-    "theme"
-  ],
-  "peerDependencies": {
-    "@astrojs/starlight": "^0.35.2",
-    "astro": "^5.13.2"
-  },
-  "dependencies": {
-    "typedoc": "^0.28.10",
-    "typedoc-plugin-markdown": "^4.8.1",
-    "unist-util-visit": "^5.0.0",
-    "yaml": "^2.8.1"
-  },
-  "devDependencies": {
-    "@astrojs/starlight": "^0.35.2",
-    "astro": "^5.13.2"
-  },
-  "type": "module",
-  "exports": {
-    ".": "./src/mod.ts",
-    "./assets/icp.svg": "./src/assets/icp.svg",
-    "./assets/layers.css": "./src/assets/layers.css",
-    "./assets/theme.css": "./src/assets/theme.css",
-    "./assets/overrides.css": "./src/assets/overrides.css",
-    "./assets/elements.css": "./src/assets/elements.css",
-    "./additional-files": "./src/additional-files.ts",
-    "./libs": "./src/libs.ts",
-    "./markdown-urls": "./src/markdown-urls.ts",
-    "./theme": "./src/theme.ts",
-    "./versioned-sidebar": "./src/versioned-sidebar/index.ts",
-    "./versioned-sidebar/components/VersionedSidebar.astro": "./src/versioned-sidebar/components/VersionedSidebar.astro",
-    "./versioned-sidebar/components/VersionDropdown.astro": "./src/versioned-sidebar/components/VersionDropdown.astro"
-  }
-}
