--- docs/plugins/src/assets/elements.css
@@ -1,28 +1,0 @@
-@layer elements {
-  /**
-   * Common
-   */
-  :root {
-    .dfn-text-highlight {
-      color: var(--sl-color-accent);
-    }
-
-    .dfn-text-gradient {
-      background-image: linear-gradient(
-        90deg,
-        var(--dfn-color-text-gradient-start),
-        var(--dfn-color-text-gradient-end)
-      );
-      -webkit-text-fill-color: transparent;
-      background-clip: text;
-    }
-
-    .dfn-bg-gradient {
-      background-image: linear-gradient(
-        0deg,
-        var(--dfn-color-bg-gradient-start),
-        var(--dfn-color-bg-gradient-end)
-      );
-    }
-  }
-}
