--- docs/plugins/src/assets/layers.css
@@ -1,1 +1,0 @@
-@layer starlight, theme, overrides, elements;
