--- docs/plugins/src/assets/overrides.css
@@ -1,9 +1,0 @@
-@layer overrides {
-  .sl-link-button {
-    border-radius: var(--dfn-border-radius);
-  }
-
-  .card {
-    border-radius: var(--dfn-border-radius);
-  }
-}
