--- docs/plugins/src/assets/theme.css
@@ -1,86 +1,0 @@
-/* Available props: https://github.com/withastro/starlight/blob/main/packages/starlight/style/props.css */
-
-@layer theme {
-  /**
-   * Common
-   */
-  :root {
-    --sl-font-system:
-      'CircularXX', ui-sans-serif, system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto,
-      'Helvetica Neue', Arial, 'Noto Sans', sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji',
-      'Segoe UI Symbol', 'Noto Color Emoji';
-
-    --dfn-color-black: #181818;
-    --dfn-color-gray-7: hsl(259, 4%, 13%);
-    --dfn-color-gray-6: hsl(259, 4%, 17%);
-    --dfn-color-gray-5: hsl(259, 4%, 20%);
-    --dfn-color-gray-4: hsl(259, 4%, 46%);
-    --dfn-color-gray-3: hsl(259, 4%, 56%);
-    --dfn-color-gray-2: hsl(259, 4%, 80%);
-    --dfn-color-gray-1: hsl(259, 4%, 93%);
-    --dfn-color-white: #ffffff;
-
-    --dfn-color-infinite: #3b00b9;
-    --dfn-color-dark-infinite: #1e005d;
-    --dfn-color-razzmatazz: #ed1e79;
-    --dfn-color-meteorite: #522785;
-    --dfn-color-picton-blue: #29abe2;
-    --dfn-color-flamingo: #f15a24;
-    --dfn-color-sea-buckthorn: #fbb03b;
-
-    --dfn-color-text-gradient-start: #6a85f1;
-    --dfn-color-text-gradient-end: #c572ef;
-
-    --dfn-border-radius: 0.75rem;
-  }
-
-  /**
-   * Dark mode
-   */
-  :root {
-    --sl-color-accent-low: var(--dfn-color-black);
-    --sl-color-accent: var(--dfn-color-picton-blue);
-    --sl-color-accent-high: var(--dfn-color-sea-buckthorn);
-
-    --sl-color-white: var(--dfn-color-white);
-    --sl-color-gray-1: var(--dfn-color-gray-1);
-    --sl-color-gray-2: var(--dfn-color-gray-2);
-    --sl-color-gray-3: var(--dfn-color-gray-3);
-    --sl-color-gray-4: var(--dfn-color-gray-4);
-    --sl-color-gray-5: var(--dfn-color-gray-5);
-    --sl-color-gray-6: var(--dfn-color-gray-6);
-    --sl-color-gray-7: var(--dfn-color-gray-7);
-    --sl-color-black: var(--dfn-color-black);
-
-    --sl-color-bg-nav: var(--dfn-color-black);
-    --sl-color-bg-sidebar: var(--dfn-color-black);
-
-    --dfn-color-bg-gradient-start: #0e031f;
-    --dfn-color-bg-gradient-end: #281447;
-  }
-
-  /**
-   * Light mode
-   */
-  :root[data-theme='light'] {
-    --sl-color-accent-low: var(--dfn-color-dark-infinite);
-    --sl-color-accent: var(--dfn-color-infinite);
-    --sl-color-accent-high: var(--dfn-color-flamingo);
-
-    --sl-color-white: var(--dfn-color-black);
-    --sl-color-gray-1: var(--dfn-color-gray-7);
-    --sl-color-gray-2: var(--dfn-color-gray-6);
-    --sl-color-gray-3: var(--dfn-color-gray-5);
-    --sl-color-gray-4: var(--dfn-color-gray-4);
-    --sl-color-gray-5: var(--dfn-color-gray-3);
-    --sl-color-gray-6: var(--dfn-color-gray-2);
-    --sl-color-gray-7: var(--dfn-color-gray-1);
-    --sl-color-black: var(--dfn-color-white);
-
-    --sl-color-bg-nav: #e8e4ed;
-    --sl-color-bg-sidebar: #f1eef5;
-
-    --dfn-color-bg-gradient-start: #c0d9ff;
-    --dfn-color-bg-gradient-end: #f0b9e5;
-  }
-}
