--- docs/plugins/src/utils/constants.ts
@@ -1,5 +1,0 @@
-import path from "node:path";
-
-export const CONTENT_DIR = path.join("src", "content");
-export const TMP_DIR = path.join(CONTENT_DIR, "tmp");
-export const DOCS_DIR = path.join(CONTENT_DIR, "docs");
