--- docs/plugins/src/versioned-sidebar/components/VersionDropdown.astro
@@ -1,100 +1,0 @@
----
-const IS_DEV = import.meta.env.DEV;
-const BASE_URL = import.meta.env.BASE_URL;
-const DOCS_VERSIONS_DROPDOWN_TITLE_VERSION = import.meta.env.DOCS_VERSIONS_DROPDOWN_TITLE_VERSION;
----
-
-<div>
-  <label for="sidebar-version-select" class="sidebar-version-select-label">
-    Version{DOCS_VERSIONS_DROPDOWN_TITLE_VERSION ? ` (${DOCS_VERSIONS_DROPDOWN_TITLE_VERSION})` : ''}
-  </label>
-  <select id="sidebar-version-select" aria-label="Select documentation version"></select>
-</div>
-
-<script define:vars={{ IS_DEV, BASE_URL }} is:inline>
-    const BASE_URL_NO_TRAILING_SLASH = BASE_URL.replace(/\/$/, '');
-
-    function removeBaseUrlFromPath(pathname) {
-        const relative = pathname.startsWith(BASE_URL)
-            ? pathname.slice(BASE_URL.length)
-            : pathname.startsWith(BASE_URL_NO_TRAILING_SLASH)
-                ? pathname.slice(BASE_URL_NO_TRAILING_SLASH.length)
-                : pathname;
-
-        return relative.startsWith('/') ? relative.slice(1) : relative;
-    }
-
-    async function initVersionDropdown() {
-        const select = document.getElementById('sidebar-version-select');
-        if (!select || !(select instanceof HTMLSelectElement)) {
-            console.error('No select element found');
-            return;
-        }
-
-        let versions = [];
-        if (!IS_DEV) {
-            try {
-                const versionsUrl = document.getElementsByName('versions-json-path')[0]?.getAttribute('content');
-                if (!versionsUrl) {
-                    console.error('No versions.json path found');
-                    return;
-                }
-                const res = await fetch(versionsUrl, { cache: 'no-store' });
-                if (res.ok) {
-                    const data = await res.json();
-                    if (Array.isArray(data)) {
-                        versions = data;
-                    }
-                }
-            } catch (e) {
-                console.error('Error fetching versions.json', e);
-            }
-        }
-
-        const currentVersion = (BASE_URL.replace(/\/$/, '').split('/').pop() || '').toLowerCase();
-        const { pathname, search, hash } = window.location;
-
-        // If the current URL's base segment doesn't match any known versions,
-        // add a synthetic option so users can see/select the current build.
-        if (!versions.some(v => v.path.toLowerCase() === currentVersion)) {
-            versions.unshift({ path: currentVersion, label: currentVersion });
-        }
-
-        select.innerHTML = '';
-        for (const v of versions) {
-            const newBase = BASE_URL.replace(/([^/]+)\/$/, `${v.path}/`);
-            const relative = removeBaseUrlFromPath(pathname);
-            const href = newBase + relative + search + hash;
-            const opt = document.createElement('option');
-            opt.value = href;
-            opt.textContent = v.label;
-            if (v.path.toLowerCase() === currentVersion) {
-                opt.selected = true;
-            }
-            select.appendChild(opt);
-        }
-
-        select.addEventListener('change', () => {
-            const href = select.value;
-            const current = pathname + search + hash;
-            if (href && href !== current) {
-                window.location.href = href;
-            }
-        });
-    };
-
-    initVersionDropdown();
-</script>
-
-<style>
-    #sidebar-version-select {
-        width: 100%;
-        padding: 0.25rem 0.5rem;
-        border-radius: 0.375rem;
-    }
-    .sidebar-version-select-label {
-        display: block;
-        font-size: var(--sl-text-sm);
-        cursor: pointer;
-    }
-</style>
