--- docs/plugins/src/versioned-sidebar/components/VersionedSidebar.astro
@@ -1,18 +1,0 @@
----
-import MobileMenuFooter from '@astrojs/starlight/components/MobileMenuFooter.astro';
-import SidebarPersister from '@astrojs/starlight/components/SidebarPersister.astro';
-import SidebarSublist from '@astrojs/starlight/components/SidebarSublist.astro';
-import VersionDropdown from './VersionDropdown.astro';
-
-const { sidebar } = Astro.locals.starlightRoute;
----
-
-<VersionDropdown />
-
-<SidebarPersister>
-	<SidebarSublist sublist={sidebar} />
-</SidebarPersister>
-
-<div class="md:sl-hidden">
-	<MobileMenuFooter />
-</div>
