--- docs/plugins/tsconfig.json
@@ -1,6 +1,0 @@
-{
-  "extends": "astro/tsconfigs/strictest",
-  "compilerOptions": {
-    "types": ["@astrojs/starlight/locals"]
-  }
-}
