+++ docs/scripts/utils/string.ts
@@ -1,5 +1,5 @@
 export function idFromFilename(fileName: string): string {
-  return fileName.replace(/\.mdx?$/, "");
+  return fileName.replace(/\.mdx?$/, '');
 }
 
 export function titleFromFilename(fileName: string): string {
@@ -7,9 +7,9 @@ export function titleFromFilename(fileName: string): string {
 }
 
 export function titleFromId(id: string): string {
-  return id.replace(/-/g, " ");
+  return id.replace(/-/g, ' ');
 }
 
 export function titleFromIdCapitalized(id: string): string {
-  return titleFromId(id).replace(/\b\w/g, (char) => char.toUpperCase());
+  return titleFromId(id).replace(/\b\w/g, char => char.toUpperCase());
 }
