--- docs/src/pages/index.astro
@@ -1,136 +1,0 @@
----
-// Import the `<StarlightPage>` component first to set up cascade layers.
-import StarlightPage from '@astrojs/starlight/components/StarlightPage.astro';
-import { Card, CardGrid, LinkButton } from '@astrojs/starlight/components';
-import icpSvg from '@dfinity/starlight/assets/icp.svg';
----
-
-<StarlightPage
-  frontmatter={{
-    title: 'ICP JavaScript SDK Core',
-    next: false,
-    prev: false,
-    tableOfContents: false,
-    hero: {
-      title: `ICP JavaScript SDK <span class="dfn-text-gradient">Core</span>`,
-      tagline:
-        'Build applications that interact with the Internet Computer from browsers, Node.js, and other JavaScript runtimes.',
-      image: {
-        file: icpSvg,
-        alt: 'Internet Computer Logo',
-      },
-    },
-    head: [
-      {
-        tag: 'style',
-        content: `
-          .sl-container {
-            max-width: 70rem;
-          }
-        `
-      }
-    ]
-  }}
->
-  <section>
-    <h3>Get Started</h3>
-
-    <CardGrid>
-      <Card title="Installation">
-        <LinkButton href="/core/latest/installation" variant="minimal" icon="right-caret" iconPlacement="start">
-          Install the <code class="dfn-text-highlight">@icp-sdk/core</code> package
-        </LinkButton>
-      </Card>
-
-      <Card title="Quick Start">
-        <LinkButton href="/core/latest/quick-start" variant="minimal" icon="right-caret" iconPlacement="start">
-          Get started with the <code class="dfn-text-highlight">@icp-sdk/core</code> package
-        </LinkButton>
-      </Card>
-    </CardGrid>
-  </section>
-
-  <section>
-    <h3>Modules</h3>
-
-    <CardGrid>
-      <Card title="Agent">
-        Build applications that interact with the Internet Computer from browsers, Node.js, and
-        other JavaScript runtimes.
-        <br />
-        <LinkButton href="/core/latest/libs/agent" variant="minimal" icon="right-caret" iconPlacement="start">
-          See <code class="dfn-text-highlight">@icp-sdk/core/agent</code> docs
-        </LinkButton>
-      </Card>
-
-      <Card title="Candid">
-        JavaScript and TypeScript module to work with Candid interfaces.
-        <br />
-        <LinkButton href="/core/latest/libs/candid" variant="minimal" icon="right-caret" iconPlacement="start">
-          See <code class="dfn-text-highlight">@icp-sdk/core/candid</code> docs
-        </LinkButton>
-      </Card>
-
-      <Card title="Identity">
-        JavaScript and TypeScript module to manage Identities and enable simple Web Authentication flows.
-        <br />
-        <LinkButton href="/core/latest/libs/identity" variant="minimal" icon="right-caret" iconPlacement="start">
-          See <code class="dfn-text-highlight">@icp-sdk/core/identity</code> docs
-        </LinkButton>
-      </Card>
-
-      <Card title="Principal">
-        JavaScript and TypeScript module to work with Internet Computer Principals.
-        <br />
-        <LinkButton
-          href="/core/latest/libs/principal"
-          variant="minimal"
-          icon="right-caret"
-          iconPlacement="start"
-        >
-          See <code class="dfn-text-highlight">@icp-sdk/core/principal</code> docs
-        </LinkButton>
-      </Card>
-
-    </CardGrid>
-  </section>
-
-  <section>
-    <h3>Upgrading</h3>
-
-    <CardGrid>
-      <Card title="ICP JS SDK Core">
-        <p>
-          If you're using <code>@dfinity/agent</code> version 3.x, you can migrate to
-          <code>@icp-sdk/core</code> with the help of our automated migration tools or manual migration
-          guide.
-        </p>
-
-        <LinkButton
-          href="/core/latest/upgrading/v4"
-          variant="minimal"
-          icon="right-caret"
-          iconPlacement="start"
-        >
-          Migrate to <code class="dfn-text-highlight">@icp-sdk/core</code>
-        </LinkButton>
-      </Card>
-
-      <Card title="Agent JS v3">
-        <p>
-          If you're still using <code>@dfinity/agent</code> version 2.x, you can migrate to 3.x with
-          the help of our migration guide.
-        </p>
-
-        <LinkButton
-          href="/core/v3.2/upgrading/v3"
-          variant="minimal"
-          icon="right-caret"
-          iconPlacement="start"
-        >
-          Migrate to <code class="dfn-text-highlight">@dfinity/agent@v3</code>
-        </LinkButton>
-      </Card>
-    </CardGrid>
-  </section>
-</StarlightPage>
