--- docs/tsconfig.libs.json
@@ -1,15 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "files": [],
-  "references": [
-    { "path": "../packages/agent" },
-    { "path": "../packages/assets" },
-    { "path": "../packages/auth-client" },
-    { "path": "../packages/candid" },
-    { "path": "../packages/core" },
-    { "path": "../packages/identity" },
-    { "path": "../packages/identity-secp256k1" },
-    { "path": "../packages/principal" },
-    { "path": "../packages/use-auth-client" }
-  ]
-}
